Add codeassist.find_definition_location
Add an interface to allow clients to find where objects are defined.

A method stub for the server command already existed - this extends it to add
functionality. One note is that the old method stub was written to return the
results of the Rope operation directly. The new method returns a dictionary of
the form:

```
{
    'result': value,
    'realpath': value,
    'path': value,
    'name': value,
    'lineno': value,
}
```