Add extra options for `rename_view`
The Rope rename command takes a few options that define how it will rename
occurrences. Expose these options in the API so the client can optionally
specify them.

Should not change how this method interacts with existing clients. The new
parameters are optional and will default to the old values.