Pass code directly in request to `code_assist_definition_view`
Previously, `code_assist_definition_view` loaded the code from disk, but this
may cause problems if the user is looking at a modified version of their file.
Add an option to pass the code directly in the request. If no code is passed, it
will then fall back to loading from disk.